+++ now_with_fractions.py
