+++ super_awesome_library.py
